# [FSDP] Correctly pin_memory for CPU offload.