Stack from [ghstack](https://github.com/ezyang/ghstack):
* __->__ #70639

x.pin_memory() results in a copied tensor in pinned memory, so it
needs to be reassigned. Internal benchmark shows ~6% win on ResNet model and ~2% win on roberta model.

Differential Revision: [D33300535](https://our.internmc.facebook.com/intern/diff/D33300535/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D33300535/)!