@rohan-varma 2022-01-08
Actually for GPTXXXL (6.7B params) before/after results are showing a regression when we pin_memory, not sure why:

Before:
train 0th step, total_latency: 55.0037734375sec, forward_time: 7.564185546875sec, backward_time: 28.72830078125sec, step_time: 18.70510546875sec, zero_grad_time: 0.00618393611907959sec, TFLOP/s/GPU: 31.596219005972518
train 1th step, total_latency: 55.9666484375sec, forward_time: 8.551041015625sec, backward_time: 28.742361328125sec, step_time: 18.66860546875sec, zero_grad_time: 0.00463974380493164sec, TFLOP/s/GPU: 31.052623664376345
train 2th step, total_latency: 55.73633203125sec, forward_time: 8.10777783203125sec, backward_time: 28.725498046875sec, step_time: 18.898326171875sec, zero_grad_time: 0.004729856014251709sec, TFLOP/s/GPU: 31.18094083966163
train 3th step, total_latency: 55.5940234375sec, forward_time: 8.18749755859375sec, backward_time: 28.70113671875sec, step_time: 18.700787109375sec, zero_grad_time: 0.004604928016662598sec, TFLOP/s/GPU: 31.26075725819595

After:

grad_time: 0.006830080032348633sec, TFLOP/s/GPU: 29.471821130339407
train 1th step, total_latency: 59.356640625sec, forward_time: 8.0635546875sec, backward_time: 32.608416015625sec, step_time: 18.6782265625sec, zero_grad_time: 0.006440959930419922sec, TFLOP/s/GPU: 29.279137993435658
train 2th step, total_latency: 59.329765625sec, forward_time: 8.12592236328125sec, backward_time: 32.565173828125sec, step_time: 18.63353515625sec, zero_grad_time: 0.0051333122253417966sec, TFLOP/s/GPU: 29.292400760029867
train 3th step, total_latency: 59.83630078125sec, forward_time: 8.245322265625sec, backward_time: 32.91680859375sec, step_time: 18.66748828125sec, zero_grad_time: 0.006681600093841553sec, TFLOP/s/GPU: 29.044430370781324